Split GUI audio settings control tests
