+++ src/gui_app/tests.rs
@@ -11,6 +11,7 @@ use radiant::{
 };
 use std::{fs, path::PathBuf, sync::mpsc};
 
+mod audio_settings_controls;
 mod shortcuts_context;
 mod status_bar;
 
@@ -688,210 +689,6 @@ fn sample_row_hit_target_survives_frame_refresh_between_press_and_release() {
         ))
     );
     assert!(!refreshed_hit_target.common().state.pressed);
-}
-
-#[test]
-fn top_status_bar_replaces_text_labels_with_volume_slider_and_audio_pill() {
-    let mut state = GuiAppState::load_default().expect("default state loads");
-    state.audio_output_resolved = Some(super::ResolvedOutput {
-        host_id: String::from("wasapi"),
-        device_name: String::from("Studio"),
-        sample_rate: 48_000,
-        buffer_size_frames: None,
-        channel_count: 2,
-        used_fallback: false,
-    });
-    let frame = radiant::runtime::UiSurface::new(super::top_status_bar(&state).into_node()).frame(
-        Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(320.0, 30.0)),
-        &radiant::theme::ThemeTokens::default(),
-    );
-    let texts = frame
-        .paint_plan
-        .primitives
-        .iter()
-        .filter_map(|primitive| match primitive {
-            PaintPrimitive::Text(text) => Some(text.text.as_str().to_string()),
-            _ => None,
-        })
-        .collect::<Vec<_>>();
-    let slider_fills = frame
-        .paint_plan
-        .primitives
-        .iter()
-        .filter_map(|primitive| match primitive {
-            PaintPrimitive::FillRect(fill)
-                if fill.widget_id == super::VOLUME_SLIDER_ID
-                    && fill.rect.width() > 0.0
-                    && fill.rect.height() > 0.0 =>
-            {
-                Some(fill)
-            }
-            _ => None,
-        })
-        .count();
-
-    assert!(!texts.iter().any(|text| text == "Wavecrate"));
-    assert!(!texts.iter().any(|text| text == "Wavecrate GUI"));
-    assert!(!texts.iter().any(|text| text == "ready"));
-    assert!(texts.iter().any(|text| text == "Audio"), "{texts:?}");
-    assert!(slider_fills >= 2, "expected track and fill rects");
-}
-
-#[test]
-fn volume_slider_drag_emits_normalized_volume() {
-    assert_eq!(
-        radiant::runtime::UiSurface::new(super::audio_settings::volume_slider(0.25).into_node(),)
-            .dispatch_widget_output(
-                super::VOLUME_SLIDER_ID,
-                radiant::widgets::WidgetOutput::typed(
-                    radiant::widgets::SliderMessage::ValueChanged { value: 0.75 },
-                ),
-            ),
-        Some(super::GuiMessage::SetVolume(0.75))
-    );
-}
-
-#[test]
-fn default_gui_volume_state_clamps() {
-    let mut state = GuiAppState::load_default().expect("default state loads");
-
-    state.set_volume(1.5);
-    assert_eq!(state.volume, 1.0);
-
-    state.set_volume(-0.5);
-    assert_eq!(state.volume, 0.0);
-}
-
-#[test]
-fn audio_engine_pill_activates_settings_toggle() {
-    let mut pill = super::audio_settings::AudioEnginePill::new(String::from("48 kHz"), false);
-    let bounds = Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(66.0, 18.0));
-    assert!(
-        pill.handle_input(
-            bounds,
-            WidgetInput::PointerPress {
-                position: Point::new(24.0, 8.0),
-                button: PointerButton::Primary,
-                modifiers: Default::default(),
-            },
-        )
-        .is_none()
-    );
-    let output = pill
-        .handle_input(
-            bounds,
-            WidgetInput::PointerRelease {
-                position: Point::new(24.0, 8.0),
-                button: PointerButton::Primary,
-                modifiers: Default::default(),
-            },
-        )
-        .expect("audio pill should activate");
-
-    assert_eq!(
-        output.typed_ref::<super::GuiMessage>(),
-        Some(&super::GuiMessage::ToggleAudioSettings)
-    );
-}
-
-#[test]
-fn audio_settings_snapshot_uses_cached_device_options() {
-    let mut state = gui_state_for_span_tests();
-    state.audio_hosts = vec![super::AudioHostSummary {
-        id: String::from("cached-host"),
-        label: String::from("Cached Host"),
-        is_default: true,
-    }];
-
-    let snapshot = super::audio_settings::AudioSettingsSnapshot::from_app_state(&state);
-
-    assert_eq!(snapshot.audio_hosts.len(), 1);
-    assert_eq!(snapshot.audio_hosts[0].id, "cached-host");
-}
-
-#[test]
-fn audio_engine_detail_distinguishes_selected_host_from_runtime_fallback() {
-    let mut state = gui_state_for_span_tests();
-    state.audio_output_config.host = Some(String::from("asio"));
-    state.audio_hosts = vec![
-        super::AudioHostSummary {
-            id: String::from("wasapi"),
-            label: String::from("WASAPI"),
-            is_default: true,
-        },
-        super::AudioHostSummary {
-            id: String::from("asio"),
-            label: String::from("ASIO"),
-            is_default: false,
-        },
-    ];
-    state.audio_output_resolved = Some(super::ResolvedOutput {
-        host_id: String::from("wasapi"),
-        device_name: String::from("Studio"),
-        sample_rate: 48_000,
-        buffer_size_frames: None,
-        channel_count: 2,
-        used_fallback: true,
-    });
-
-    assert_eq!(
-        state.audio_engine_detail_label(),
-        "ASIO selected | using WASAPI | Studio | 48 kHz"
-    );
-}
-
-#[test]
-fn audio_sample_rate_label_matches_status_chip_format() {
-    assert_eq!(super::format_sample_rate_label(48_000), "48 kHz");
-    assert_eq!(super::format_sample_rate_label(44_100), "44.1 kHz");
-    assert_eq!(super::format_sample_rate_label(960), "960 Hz");
-}
-
-#[test]
-fn audio_settings_popover_stays_output_only() {
-    let mut state = GuiAppState::load_default().expect("default state loads");
-    state.audio_settings_error = None;
-    state.audio_hosts = vec![super::AudioHostSummary {
-        id: String::from("asio"),
-        label: String::from("ASIO"),
-        is_default: false,
-    }];
-    state.audio_devices = vec![super::AudioDeviceSummary {
-        host_id: String::from("asio"),
-        name: String::from("Studio Out"),
-        is_default: true,
-    }];
-    state.audio_sample_rates = vec![44_100, 48_000];
-    let frame = radiant::runtime::UiSurface::new(super::audio_settings_popover(&state).into_node())
-        .frame(
-            Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(480.0, 360.0)),
-            &radiant::theme::ThemeTokens::default(),
-        );
-    let texts = frame
-        .paint_plan
-        .primitives
-        .iter()
-        .filter_map(|primitive| match primitive {
-            PaintPrimitive::Text(text) => Some(text.text.as_str().to_string()),
-            _ => None,
-        })
-        .collect::<Vec<_>>();
-
-    assert!(
-        !texts.iter().any(|text| text == "Audio Engine"),
-        "{texts:?}"
-    );
-    assert!(texts.iter().any(|text| text == "Backend"), "{texts:?}");
-    assert!(texts.iter().any(|text| text == "Output"), "{texts:?}");
-    assert!(texts.iter().any(|text| text == "Sample Rate"), "{texts:?}");
-    assert!(
-        texts.iter().any(|text| text == "Clear Rebuildable Caches"),
-        "{texts:?}"
-    );
-    assert!(
-        !texts.iter().any(|text| text.contains("Input")),
-        "{texts:?}"
-    );
 }
 
 #[test]
